pyproject.toml
[tool.poetry]
name = "seed"
version = "0.1.0"
description = ""
authors = ["Ivan Moskovchenko <imosk72@yandex-team.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
pyyaml = "^6.0.2"
requests = "^2.32.3"
argparse = "^1.4.0"

#[tool.poetry.group.learner.dependencies]



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
